Factor out server response pump
